requirements.txt
django==3.2
djangorestframework
django-finalware
django-cors-headers
mysqlclient
pyyaml
drf-yasg
django-filter
django-spaghetti-and-meatballs
django-decorator-include
pymemcache
django-storages
django-mysql
celery==4.4.7
django-celery-beat
django_celery_results
flower==0.9.7
pyjwkest
jsonfield
django-threadlocals
more_itertools
joblib
